[IMP] sale_order_product_picker: Changes to improve usability of +1
Added changes to make the button +1 faster. With these changes, what we
do is to store the changes to be made while pressing the +1 button. Each
time the button is pressed, the timer that will execute the writing of
the lines is reset. The time to write the pending lines is 1 sec by
default, but it can be increased using the system parameter
sale_order_product_picker.delay and the number of seconds to wait.

This relieves the server of unnecessary calls, since everything will be
written at the same time. And the user will not have the feeling of
slowness when writing on 3G networks.